2009-06-04 Roman - !!:/DemoDB: DB vcaBase.db, vcaTest.db and vcaAGLKS.db translation to multilanguage (Russian and English) is finished. - !!:/SYS: Demo config file translation to multilanguage (Russian, English and Ukrainian) is finished. - FIX:/Archive: Function TArchive::getVal() is fixed for buffer's data request. - FIX:/Archive/DBArch: Limited time variables is fixed into function ModVArchEl::getValProc(). - !!:/SYS: Multilanguage text variables support is added to subsystem "Archive" and "DB". - !!:/DAQ/LogicLev: Multilanguage text variables into template's variable support is added. - !!:/UI/VCAEngine: Multilanguage text variables support is added. - FIX:/UI/VCAEngine: Process edit tab into configurators for widgets is fixed for child's attributes edit support and modify flag for widget setting. - FIX:/UI/Vision: Setting the modify flag for now changed attribute into the attributes inspector is added.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@900 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/moduls/arhiv/DBArch/val.cpp
@@ -157,8 +157,8 @@ void ModVArchEl::getValsProc( TValBuf &buf, long long ibegIn, long long iendIn )
     iendIn = (iendIn/period())*period();
 
     //> Prepare border
-    long ibeg = vmax( ibegIn, begin() );
-    long iend = vmin( iendIn, end() );
+    long long ibeg = vmax( ibegIn, begin() );
+    long long iend = vmin( iendIn, end() );
 
     if( iend < ibeg )	return;
 
@@ -214,7 +214,7 @@ void ModVArchEl::getValsProc( TValBuf &buf, long long ibegIn, long long iendIn )
 
 TVariant ModVArchEl::getValProc( long long *tm, bool up_ord )
 {
-    long long itm = tm?*tm:SYS->curTime();
+    long long itm = tm ? *tm : SYS->curTime();
     itm = (itm/period())*period()+((up_ord && itm%period())?period():0);
 
     TConfig cf(NULL);
